delete engine lineage parse
